powerpc sstep: Add tests for Prefixed Add Immediate
Use the support for testing compute type instructions to test Prefixed
Add Immediate (paddi). The patch_site macro is used for testing compute
type instructions. This allows for patching in a single word. As
prefixed instructions need two words, add a second patch site directly
afterwards. As the instruction that is used for patching is a nop this
will not effect the other single word instructions.

Skip the paddi tests if ISA v3.1 is unsupported.

Signed-off-by: Jordan Niethe <jpn@ozlabs.au.ibm.com>
Signed-off-by: Balamuruhan S <bala24@linux.ibm.com>