powerpc/sstep: rfc to handle dword storage in fp_state/vr_state reg for lxvd2x
In emulate_vsx_{load/store}(), for lxvd2x/stxvd2x instruction in LE system
and BE on LE system 1st 8byte data is written to dword1 (2nd 8byte) of
vsx_reg and 2nd 8byte data is written to dword0 (1st 8byte) of vsx_reg.
If MSR_FP/MSR_VR is set load_vsrn() performs lxvd2x and xxswapd to load
in vsx register dword appropriately, but if MSR_FP/MSR_VR is not set,
then the datas' dword storage is not handled to be saved in
fp_state.fpr/vr_state.vr appropriately. This patch tries to handle them.

Signed-off-by: Balamuruhan S <bala24@linux.ibm.com>